+++ CHANGELOG.md
@@ -30,9 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Breaking: In the `Source` trait, the method `current_frame_len()` was renamed to `current_span_len()`.
 - Breaking: `Decoder` now outputs `f32` samples.
 - Breaking: The term 'frame' was renamed to 'span' in the crate and documentation.
-- Breaking: Sources now use `f32` samples. To convert to and from other types of samples use functions from
-  `dasp_sample` crate. For example `DaspSample::from_sample(sample)`. Remove `integer-decoder` feature.
-
+- Breaking: `LoopedDecoder` now returns `None` if seeking fails during loop reset
+- Breaking: Sources now use `f32` samples. To convert to and from other types of samples use
+            functions from `dasp_sample` crate. For example `DaspSample::from_sample(sample)`.
 
 ### Fixed
 - `ChannelVolume` no longer clips/overflows when converting from many channels to
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - WAV and FLAC decoder duration calculation now calculated once and handles very large files
   correctly
 - Removed unwrap() calls in MP3, WAV, FLAC and Vorbis format detection for better error handling
+- `LoopedDecoder::size_hint` now correctly indicates an infinite stream
 
 ### Deprecated
 - Deprecated `Sample::zero_value()` function in favor of `Sample::ZERO_VALUE` constant